+++ Assets/Animations/Pedestrians/Beetles/Beetle Crash L.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a48c0a646d60645859ab6a4565ed7bc6
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
