+++ Assets/Animations/Pedestrians/Beetles/Beetle Crash R.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9a66cbcd145fe42628ea22b7ae29cb19
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
