+++ Assets/Prefabs/Game Prefabs/Pedestrian.prefab
@@ -55,10 +55,10 @@ CapsuleCollider:
   m_ProvidesContacts: 0
   m_Enabled: 1
   serializedVersion: 2
-  m_Radius: 0.4267161
-  m_Height: 1.7325692
+  m_Radius: 0.3197279
+  m_Height: 0.968702
   m_Direction: 1
-  m_Center: {x: 0.02286905, y: -0.02970624, z: 0.08173883}
+  m_Center: {x: 0.12985724, y: -0.030792922, z: 0.08173883}
 --- !u!54 &3796284116853647316
 Rigidbody:
   m_ObjectHideFlags: 0
@@ -147,7 +147,7 @@ Transform:
   serializedVersion: 2
   m_LocalRotation: {x: 0.23911765, y: 0.3696438, z: -0.09904577, w: 0.89239913}
   m_LocalPosition: {x: 0, y: 0, z: 0}
-  m_LocalScale: {x: 0.1, y: 0.1, z: 0.1}
+  m_LocalScale: {x: 0.06, y: 0.06, z: 0.06}
   m_ConstrainProportionsScale: 1
   m_Children: []
   m_Father: {fileID: 8182307269469577234}
