--- Assets/Scripts/Pedestrian/PedestrianNavigationController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: db443f28849774794b1349acb4804a2c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
